Changes to support pre-cloned catalog present on disk.
The service now first checks if the catalog by its name and repoURL
already exists under ./DATA/ folder. If it does then it just pulls
updates and serves the catalog without cloning newly. In case the pull
fails then whatever is on the disk is served.

If the catalogName or repoURL do not match to existing catalog, then
that catalog is newly cloned.